refactor(utils): share raw response parse helpers
Move strict numeric parsing, HTML attribute extraction, and blank response detection into a lightweight utils header for reuse across broker parsers.

Keep IntradeBar-specific active-trade parsing logic inside the platform parser.
